moved pagination component to shared dir
